fixed linters errors
