+++ src/pages/homepage/home.js
@@ -1,0 +1,26 @@
+import React from 'react';
+
+const Home = () => (
+  <div>
+    <h2>Welcome to our page</h2>
+    <div className="content">
+      The math magician calculator is a calcuator app for users to do calcuations of any sort.
+      The simple
+      calculor apps performs all kinds of complex calcuations that the normal
+      calculations can do.
+      <br />
+      Its builts with the maths logic in mind to enable fill your calculation.
+      Some of the functionalities of the calculor apps are not but limited to
+      <ul>
+        <li>Simple calculating function</li>
+        <li>Calcuate all forms of figures</li>
+        <li>Used by many for complex maths</li>
+      </ul>
+
+      The calculator app is bult with Reactjs, which allows you to manage both class based and
+      functional component. Its ensures that your functional components uses hooks to manage state.
+    </div>
+  </div>
+);
+
+export default Home;
