+++ src/pages/quote/quote.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Quote = () => (
+  <div>
+    There is like ten minutes when it is like, Okay, wait, who is this guy again?
+    And then, you know, I just put on the calculator watch and the glasses, and just be all,
+    you know, inappropriate. And then it just works out fine.
+  </div>
+);
+
+export default Quote;
