--- src/pages/quote/quote.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Quote = () => {
-    return (
-        <div>
-            There's like ten minutes when it's like, 'Okay, wait, who is this guy again?' And then, you know, I just put on the calculator watch and the glasses, and just be all, you know, inappropriate. And then it just works out fine.
-        </div>
-    )
-}
-
-export default Quote;
